+++ pkg/querier/error_translate_queryable_test.go
@@ -202,6 +202,10 @@ func (t errorTestQuerier) LabelValues(context.Context, string, ...*labels.Matche
 	return nil, nil, t.err
 }
 
+func (t errorTestQuerier) LabelValuesStream(context.Context, string, ...*labels.Matcher) storage.LabelValues {
+	return storage.ErrLabelValues(t.err)
+}
+
 func (t errorTestQuerier) LabelNames(context.Context, ...*labels.Matcher) ([]string, annotations.Annotations, error) {
 	return nil, nil, t.err
 }